WIP make sure current_user is the owner of the the bike_event to see remove marker link

+++ app/controllers/bike_events_controller.rb
@@ -24,10 +24,12 @@ class BikeEventsController < ApplicationController
   end
 
   def destroy
-    bike_event = BikeEvent.find(params[:id])
-    bike_event.destroy
-    head :no_content  
-    # sends a 204
+    if current_user.id == bike_event.user_id
+      bike_event = BikeEvent.find(params[:id])
+      bike_event.destroy
+      head :no_content  
+      # sends a 204
+    end
   end
 
   private
